:neckbeard: Use JSON payload, reduce checks in API
Some of the checks within the this app are doubling up the validation
that takes place within the Akamai API. Some of these checks are better
placed within this app as additional information is able to be provided
to the client, plus it prevents a request to the Akamai API when we know
that is will result in a failure. However, some of the checks e.g. the
size of the request has been left to happen at the Akamai end as this is
a value that might change and we don't want to be in a position where
this app also needs to be updated (not when there isn't a business rule
that we wish to apply).