:fire: Azure Subscription can not be a variable
Ideally the name of the service connection aka the subscription (as
specified within the pipeline/project in Azure DevOps) would be able to
be specified with a variable. The reason this would be useful is the
subscription varies by environment and it could have been set in Vault.
However, it appears that as part of the template validation the name of
the service connection is evaluated to check there is available.
However, having the service connection not exist until runtime means the
literal value of the supposed variable is used and isn't going to be
found, thus failing the pipeline validation.